+++ src/main/java/roomescape/auth/interceptor/AdminPageAuthorizationInterceptor.java
@@ -5,7 +5,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import java.util.Optional;
 import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.auth.AuthorizationExtractor;
-import roomescape.auth.Role;
+import roomescape.domain.Role;
 import roomescape.infrastructure.JwtTokenProvider;
 
 public class AdminPageAuthorizationInterceptor implements HandlerInterceptor {
