+++ src/main/java/roomescape/domain/Member.java
@@ -7,7 +7,6 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import roomescape.auth.Role;
 import roomescape.exception.LoginFailedException;
 
 @Entity
