+++ src/main/java/roomescape/dto/request/MemberRequest.java
@@ -2,8 +2,8 @@ package roomescape.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
-import roomescape.auth.Role;
 import roomescape.domain.Member;
+import roomescape.domain.Role;
 
 public record MemberRequest(
         @Email(message = "잘못된 이메일 형식입니다.") String email,
