+++ src/test/java/roomescape/unit/domain/MemberTest.java
@@ -2,8 +2,8 @@ package roomescape.unit.domain;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
-import roomescape.auth.Role;
 import roomescape.domain.Member;
+import roomescape.domain.Role;
 import roomescape.exception.LoginFailedException;
 
 class MemberTest {
