+++ src/test/java/roomescape/unit/fake/FakeMemberRepositoryTest.java
@@ -5,8 +5,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.List;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
-import roomescape.auth.Role;
 import roomescape.domain.Member;
+import roomescape.domain.Role;
 
 class FakeMemberRepositoryTest {
