+++ src/test/java/roomescape/unit/fake/FakeReservationRepositoryTest.java
@@ -7,10 +7,10 @@ import java.time.LocalTime;
 import java.util.List;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
-import roomescape.auth.Role;
 import roomescape.domain.Member;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
+import roomescape.domain.Role;
 import roomescape.domain.Theme;
 import roomescape.domain.repository.ReservationRepository;
