+++ src/test/java/roomescape/unit/presentation/AdminControllerTest.java
@@ -19,7 +19,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import roomescape.auth.AuthorizationExtractor;
-import roomescape.auth.Role;
+import roomescape.domain.Role;
 import roomescape.dto.request.AdminReservationCreateRequest;
 import roomescape.dto.response.ReservationResponse;
 import roomescape.dto.response.ReservationTimeResponse;
