+++ src/test/java/roomescape/unit/service/MemberServiceTest.java
@@ -3,8 +3,8 @@ package roomescape.unit.service;
 import java.util.List;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
-import roomescape.auth.Role;
 import roomescape.domain.Member;
+import roomescape.domain.Role;
 import roomescape.domain.repository.MemberRepository;
 import roomescape.dto.response.MemberResponse;
 import roomescape.service.MemberService;
